Transition to AndroidX, many bugfixes
With this commit I updated the app to use AndroidX instead of the
compatibility library. Also numerous bug fixes were introduced like
the LinesFragment now correctly saves its state, times are now rounded
based on the amount of passed seconds instead of just taking the integer
portion. When the dark theme is used the Map is also dark themed. Plenty
of small bugfixes.